Build and deploy website via GitHub pages

+++ ts/backend/script-clear-json.ts
@@ -5,14 +5,17 @@ import { Log, PATH_JSON } from './general.js';
 function main() {
     try {
         // delete json directory, and create a fresh one
-        Log.action(`Deleting '${PATH_JSON}'`);
-        fs.rmSync(PATH_JSON, { recursive: true });
-        Log.action(`Creating '${PATH_JSON}'`);
-        fs.mkdirSync(PATH_JSON);
-        Log.success('Done');
+        if (fs.existsSync(PATH_JSON)) {
+            Log.action(`Deleting '${PATH_JSON}'`, () => {
+                fs.rmSync(PATH_JSON, { recursive: true });
+            });
+        }
+        Log.action(`Creating '${PATH_JSON}'`, () => {
+            fs.mkdirSync(PATH_JSON);
+        });
     }
     catch (err: any) {
-        Log.error(err.stack);
+        Log.error(`${err.stack}`);
         process.exit(1);
     }
 }
